requirements.txt
google-cloud-storage==1.39.0
googleapis-common-protos==1.53.0
Flask==1.1.1
Flask-RESTful==0.3.9
google-cloud-logging
joblib
xgboost